Block View with Sidebar
Block View has been redone to work with the sidebar. I am having continued issues adding a hover-over listener that will store the target semester node during the middle of the drag & drop from the sidebar.

In this version, you can select/highlight a target semester by double clicking, and then dragging a class node from the side bar will drop it into the desired semester. I will continue to tinker with this, there seems to be promise with storing the hovered semester as a useState().